rename: hooks 폴더 안에 auth 폴더 생성하여 관련 파일 이동 적용
